requirements.txt
numpy<2
torch>=2.4
onnx
opencv-python
streamlit
transformers
timm
einops
ultralytics>=8.4.19
tqdm
pillow
scipy
google-generativeai>=0.9.0
python-dotenv>=1.0.1
